feat: command delete variables
